+++ CHANGELOG.md
@@ -3,6 +3,12 @@ ChangeLog
 
 Please also read the [Upgrade Guide](https://github.com/katzer/cordova-plugin-local-notifications/wiki/Upgrade-Guide) for more information.
 
+#### Version 0.10.3
+- fixed iOS notifications show/hide in foreground mode logic
+
+#### Version 0.10.2
+- fixed Android 8 notifications sound on the lock screen
+
 #### Version 0.10.0 
 - fixed Android 8 notifications sound